Refactor BusinessDTO and move db logic to orm package
