--- Contam/src/main/java/com/bugbusters/contam/business/Business.java
@@ -1,91 +1,0 @@
-package com.bugbusters.contam.business;
-
-/**
- *
- * @author Andreas
- */
-public class Business {
-
-    private int id;
-    private String name;
-    private String address;
-    private String description;
-    private double longitude;
-    private double latitude;
-    private int postalCode;
-
-    public Business() {
-        
-    }
-    
-    public Business(int id, String name, double longitude, double latitude) {
-        this.id = id;
-        this.name = name;
-        this.longitude = longitude;
-        this.latitude = latitude;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public int getID() {
-        return id;
-    }
-
-    public void setID(int id) {
-        this.id = id;
-    }
-
-    public double getLongitude() {
-        return longitude;
-    }
-
-    public void setLongitude(double longitude) {
-        this.longitude = longitude;
-    }
-
-    public double getLatitude() {
-        return latitude;
-    }
-
-    public void setLatitude(double latitude) {
-        this.latitude = latitude;
-    }
-
-    public int getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(int postalCode) {
-        this.postalCode = postalCode;
-    }
-}
