+++ Contam/src/main/webapp/index.jsp
@@ -1,5 +1,5 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
-<%@ page import="com.bugbusters.contam.business.Business" %>
+<%@ page import="com.bugbusters.contam.orm.business.BusinessDTO" %>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ taglib prefix="fn" uri="http://java.sun.com/jsp/jstl/functions" %>
